dev/test/com/ericsson/eniq/repository/UpdateDBUsersTest.java
package com.ericsson.eniq.repository;
import static org.junit.Assert.*;
import org.junit.Test;

import com.distocraft.dc5000.etl.rock.Meta_databases;

import ssc.rockfactory.RockException;
import ssc.rockfactory.RockFactory;

import java.io.*;
import java.sql.SQLException;
import java.sql.Connection;
public class UpdateDBUsersTest {
	
	
	 @Test
	    public void testUpdateMetaDatabasesWithValidInputs() {
	        
	        String user = "testUser";
	        String oldPassword = "oldPassword";
	        String newPassword = "newPassword";

	        
	        int result = UpdateDBUsersTest.updateMetaDatabases(user, oldPassword, newPassword);

	       
	        assertEquals(0, result); 
	    }

	    @Test
	    public void testUpdateMetaDatabasesWithNullInputs() {
	        
	        String user = null;
	        String oldPassword = null;
	        String newPassword = null;

	       
	        int result = UpdateDBUsersTest.updateMetaDatabases(user, oldPassword, newPassword);

	       
	       
	    }
	@Test
	public void testGetRockFactoryObject() throws IOException, SQLException, RockException {
		System.setProperty("CONF_DIR", "test_conf");
		//ssc.rockfactory.RockFactory factory = UpdateDBUsers.getRockFactoryObject();
	}
	
	abstract class CustomConnection implements Connection {
		private boolean closed = false;
		
		@Override
		public void close() throws SQLException
		{
			if (closed)
			{
				throw new SQLException("Connection is already closed.");}
			closed = true;
			}
		
	
		}
	
	@Test
    public void testSetEncryptionFlag() {
        
       

        
        MetaDatabaseObject metaDatabaseObject = new MetaDatabaseObject();

        
        metaDatabaseObject.setEncryption_flag("YY");

        
        assertEquals("YY", metaDatabaseObject.getEncryption_flag());
    }

    
    private static class MetaDatabaseObject {
        private String encryption_flag;

        public String getEncryption_flag() {
            return encryption_flag;
        }

        public void setEncryption_flag(String encryption_flag) {
            this.encryption_flag = encryption_flag;
        }
    }


	
@Test
public void testFinallyBlockWithNullEtlRep() {
	UpdateDBUsersTest.finallyBlock(null);
	
}
@Test
public void testUpdateMetaDatabasesWithNullValues()
{
	String user = null;
	String oldPassword = null;
	String newPassword = null;
	int result = UpdateDBUsersTest.updateMetaDatabases(user, oldPassword, newPassword);
	assertEquals(0, result);
	
}

@Test 
public void testUpdateMetaDatabasesWithNonNullValues() {
	String user = "testUser";
	String oldPassword = "oldPassword";
	String newPassword = "newPassword";
	int result = UpdateDBUsersTest.updateMetaDatabases(user, oldPassword, newPassword);
}
@Test


public void testSetEncryptionFlag1() {
    
    Meta_databases metaDatabaseObject = new Meta_databases(null, false);

   
    metaDatabaseObject.setEncryption_flag("YY");
      String S = "YY";
    
    assertEquals(S, metaDatabaseObject.getEncryption_flag());


}    @Test
public void testGetRockFactoryObject1() {
    System.setProperty("CONF_DIR", "path/to/your/test/conf"); 
	RockFactory rockFactory = UpdateDBUsersTest.getRockFactoryObject();

	
	assertNotNull(rockFactory);
}

private static RockFactory getRockFactoryObject() {
	
	return null;
}
private static int updateMetaDatabases(String user, String oldPassword, String newPassword) {
	
	return 0;
	//dummy push
}
private static void finallyBlock(Object object) {
	
	
	
	
}
	
}
